+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 3.1.40
+- Added Wrestling Empire support
+- GodotML setup script is now called for Dome Keeper
+
 ### 3.1.39
 - Added support for more games:
   - RUMBLE